wip(clippy): mechanical fixes for lsp-rs/core test-target lint debt
Partial fixes toward the meta-shard clippy_full gate: question_mark
rewrites in method_direction.rs, expect-free marker handling in the
path_tiny test, needless-borrow and unused-import/mut cleanups. The
remaining tail is dead-code judgment calls in the protocol inventory
machinery (error_inventory, final_surface_inventory) - catalogued for a
fresh-context pass. See #14863.

+++ crates/perl-lsp-rs-core/src/protocol/document_version.rs
@@ -486,7 +486,7 @@ mod tests {
     fn u64_above_i64_max() {
         // Construct a u64 value above i64::MAX directly in the JSON value.
         // serde_json stores this as N::PosInt(u64) with value > i64::MAX.
-        let big: u64 = u64::from(i64::MAX as u64) + 1; // i64::MAX + 1
+        let big: u64 = (i64::MAX as u64) + 1; // i64::MAX + 1
         let value = serde_json::Value::Number(serde_json::Number::from(big));
         let params = json!({"textDocument": {"version": value}});
         assert_eq!(
